[BA-40] feat: Set clickable action for logout buttons
