+++ app/src/main/java/com/example/bookingapp/pages/hotelier/ModProfilePage.kt
@@ -53,6 +53,7 @@ import com.example.bookingapp.core.ui.theme.OrangePrimary
 import com.example.bookingapp.mock_data.HotelData
 import com.example.bookingapp.models.Hotel
 import com.example.bookingapp.navigation.RootScreen
+import com.example.bookingapp.view_models.MainViewModel
 
 
 @Composable
@@ -356,7 +357,7 @@ fun Logout(navController: NavController) {
         modifier = Modifier
             .fillMaxWidth()
             .padding(vertical = 8.dp, horizontal = 16.dp)
-            .clickable { navController.navigate(RootScreen.Login.route) },
+            .clickable { MainViewModel.authViewModel.logout() },
         horizontalArrangement = Arrangement.SpaceBetween,
         verticalAlignment = Alignment.CenterVertically
     ) {
