+++ app/src/main/java/com/example/bookingapp/view_models/AuthViewModel.kt
@@ -11,6 +11,7 @@ import androidx.lifecycle.ViewModel
 import com.example.bookingapp.MainActivity
 import com.example.bookingapp.models.Account
 import com.example.bookingapp.services.AccountService
+import com.example.bookingapp.services.RetrofitClient
 import com.google.gson.Gson
 
 class AuthViewModel: ViewModel() {
@@ -39,6 +40,19 @@ class AuthViewModel: ViewModel() {
         }
     }
 
+    fun logout() {
+        try {
+            isAuthenticated = false
+            authToken = ""
+            account = Account()
+
+            clearAccount()
+            RetrofitClient.clearAuthToken()
+        } catch (e: Exception) {
+            throw Exception("Auth view model: ${e.message}")
+        }
+    }
+
     suspend fun login(email: String, password: String): Boolean {
         try {
             AccountService.login(email, password).let {
@@ -46,6 +60,9 @@ class AuthViewModel: ViewModel() {
                     account = it.first
                     authToken = it.second
                     isAuthenticated = true
+
+                    saveAccount(account, authToken)
+                    RetrofitClient.setAuthToken(authToken)
                 }
             }
         } catch (e: Exception) {
@@ -53,5 +70,34 @@ class AuthViewModel: ViewModel() {
         }
 
         return isAuthenticated
+    }
+
+    private fun saveAccount(account: Account, token: String) {
+        val accountJson = Gson().toJson(account)
+        val context = MainActivity.context
+
+        val accountPref = context.getSharedPreferences("account", Context.MODE_PRIVATE)
+        val accountEditor = accountPref.edit()
+        accountEditor.putString("account", accountJson)
+        accountEditor.apply()
+
+        val tokenPref = context.getSharedPreferences("token", Context.MODE_PRIVATE)
+        val tokenEditor = tokenPref.edit()
+        tokenEditor.putString("token", token)
+        tokenEditor.apply()
+    }
+
+    private fun clearAccount() {
+        val context = MainActivity.context
+
+        val accountPref = context.getSharedPreferences("account", Context.MODE_PRIVATE)
+        val accountEditor = accountPref.edit()
+        accountEditor.clear()
+        accountEditor.apply()
+
+        val tokenPref = context.getSharedPreferences("token", Context.MODE_PRIVATE)
+        val tokenEditor = tokenPref.edit()
+        tokenEditor.clear()
+        tokenEditor.apply()
     }
 }
